requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
aiomysql==0.2.0
cryptography==42.0.0
